+++ 3rd-party/dparser/write_tables.h
@@ -6,4 +6,4 @@ int write_c_tables(Grammar *g);
 int write_binary_tables(Grammar *g);
 int write_binary_tables_to_file(Grammar *g, FILE *fp);
 int write_binary_tables_to_string(Grammar *g, 
-				  unsigned char **str, unsigned int *str_len);
+				  unsigned char **str, size_t *str_len);
